Random_Forest Classifier  algorithm is used for train the data
Description:
This dataset contains information about 686 songs, including their metadata, audio features, and associated moods. It can be used for tasks such as music mood classification, recommendation systems, and audio analysis.

Columns:
name - Name of the song
album - Album in which the song was released
artist - Artist or band who performed the song
id - Unique identifier for the song
release_date - Release date of the song
popularity - Popularity score (integer)
length - Length of the song in milliseconds
danceability - Measure of how suitable a song is for dancing (0.0 to 1.0)
acousticness - Measure of how acoustic a song is (0.0 to 1.0)
energy - Represents the intensity and activity of a song (0.0 to 1.0)
instrumentalness - Likelihood that a song has no vocals (0.0 to 1.0)
liveness - Detects the presence of a live audience (0.0 to 1.0)
valence - Describes the musical positivity (0.0 to 1.0)
loudness - Overall loudness of the track (measured in dB)
speechiness - Detects spoken words in a track (0.0 to 1.0)
tempo - Estimated tempo of the song in BPM (beats per minute)
key - The key in which the song is played (0-11)
time_signature - Time signature of the song (number of beats per bar)
mood - Categorical label indicating the mood of the song (e.g., Happy, Sad, Energetic, Calm)
Dataset contains both categorical and continuous data in features so pre-processing technique is used and labelled data is categorical so classification algorithms can be used.